+++ src/models.rs
@@ -10,7 +10,7 @@ use rocket::serde::{Deserialize, Serialize};
 #[serde(crate = "rocket::serde")]
 pub struct User {
     #[serde(default)]
-    pub id: i32,
+    pub id: i64,
     pub username: String,
     pub password: String,
     pub email_address: Option<String>,
@@ -21,7 +21,7 @@ pub struct User {
 #[derive(Queryable, Debug, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Category {
-    pub id: i32,
+    pub id: i64,
     pub name: String,
     pub description: Option<String>,
     pub created: NaiveDateTime,
@@ -38,21 +38,21 @@ pub struct CreateCategory {
 #[derive(Queryable, Debug, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Ticket {
-    pub id: i32,
-    pub category_id: Option<i32>,
+    pub id: i64,
+    pub category_id: Option<i64>,
     pub subject: String,
     pub description: Option<String>,
     pub status: String,
     pub priority: String,
     pub created: NaiveDateTime,
-    pub reporter_id: i32,
+    pub reporter_id: i64,
 }
 
 #[derive(Deserialize, Insertable)]
 #[table_name = "tickets"]
 #[serde(crate = "rocket::serde")]
 pub struct CreateTicket {
-    pub category_id: Option<i32>,
+    pub category_id: Option<i64>,
     pub subject: String,
     pub description: Option<String>,
     pub status: String,
@@ -62,8 +62,8 @@ pub struct CreateTicket {
 #[derive(Queryable, Debug, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Assignment {
-    pub ticket_id: i32,
-    pub user_id: i32,
+    pub ticket_id: i64,
+    pub user_id: i64,
     pub assigned: NaiveDateTime,
 }
 
@@ -71,15 +71,15 @@ pub struct Assignment {
 #[table_name = "assignments"]
 #[serde(crate = "rocket::serde")]
 pub struct CreateAssignmentWithUserId {
-    pub user_id: i32,
+    pub user_id: i64,
 }
 
 #[derive(Queryable, Debug, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 pub struct Comment {
-    pub id: i32,
-    pub ticket_id: i32,
-    pub commenter_id: i32,
+    pub id: i64,
+    pub ticket_id: i64,
+    pub commenter_id: i64,
     pub content: String,
     pub commented: NaiveDateTime,
 }
@@ -88,7 +88,7 @@ pub struct Comment {
 #[table_name = "comments"]
 #[serde(crate = "rocket::serde")]
 pub struct CreateComment {
-    pub commenter_id: i32,
+    pub commenter_id: i64,
     pub content: String,
 }
 
@@ -96,7 +96,7 @@ pub struct CreateComment {
 #[serde(crate = "rocket::serde")]
 pub struct Session {
     pub uuid: String,
-    pub user_id: i32,
+    pub user_id: i64,
     pub created: NaiveDateTime,
     pub expiration: NaiveDateTime,
 }
